+++ google/cloud/spanner_dbapi/connection.py
@@ -4,14 +4,17 @@
 # license that can be found in the LICENSE file or at
 # https://developers.google.com/open-source/licenses/bsd
 
-from collections import namedtuple
+"""DB-API Connection for the Google Cloud Spanner."""
 
-from google.cloud import spanner_v1 as spanner
+# from google.cloud import spanner_v1
 
 from .cursor import Cursor
 from .exceptions import InterfaceError
 
-ColumnDetails = namedtuple("column_details", ["null_ok", "spanner_type"])
+# from .version import google_client_info
+# from google.cloud.spanner_dbapi import Cursor
+# from google.cloud.spanner_dbapi import google_client_info
+# from google.cloud.spanner_dbapi import InterfaceError
 
 
 class Connection(object):
@@ -23,117 +26,32 @@ class Connection(object):
 
     def __init__(self, database):
         self._database = database
-        self._ddl_statements = []
-
+        self.ddl_statements = []
         self.is_closed = False
 
-    def cursor(self):
-        """Factory to create a DB-API Cursor."""
-        self._raise_if_closed()
-
-        return Cursor(self)
+    @property
+    def database(self):
+        return self._database
 
     def _raise_if_closed(self):
-        """Raise an exception if this connection is closed.
+        """Helper to check the connection state before running a query.
+        Raises an exception if this connection is closed.
 
-        Helper to check the connection state before
-        running a SQL/DDL/DML query.
-
-        :raises: :class:`InterfaceError` if this connection is closed.
+        :raises: :class:`InterfaceError`: if this connection is closed.
         """
         if self.is_closed:
             raise InterfaceError("connection is already closed")
 
-    def __handle_update_ddl(self, ddl_statements):
-        """
-        Run the list of Data Definition Language (DDL) statements on the underlying
-        database. Each DDL statement MUST NOT contain a semicolon.
-        Args:
-            ddl_statements: a list of DDL statements, each without a semicolon.
-        Returns:
-            google.api_core.operation.Operation.result()
-        """
-        self._raise_if_closed()
-        # Synchronously wait on the operation's completion.
-        return self._database.update_ddl(ddl_statements).result()
-
-    def read_snapshot(self):
-        self._raise_if_closed()
-        return self._database.snapshot()
-
-    def in_transaction(self, fn, *args, **kwargs):
-        self._raise_if_closed()
-        return self._database.run_in_transaction(fn, *args, **kwargs)
-
-    def append_ddl_statement(self, ddl_statement):
-        self._raise_if_closed()
-        self._ddl_statements.append(ddl_statement)
-
-    def run_prior_DDL_statements(self):
-        self._raise_if_closed()
-
-        if not self._ddl_statements:
-            return
-
-        ddl_statements = self._ddl_statements
-        self._ddl_statements = []
-
-        return self.__handle_update_ddl(ddl_statements)
-
-    def list_tables(self):
-        return self.run_sql_in_snapshot(
-            """
-            SELECT
-              t.table_name
-            FROM
-              information_schema.tables AS t
-            WHERE
-              t.table_catalog = '' and t.table_schema = ''
-            """
-        )
-
-    def run_sql_in_snapshot(self, sql, params=None, param_types=None):
-        # Some SQL e.g. for INFORMATION_SCHEMA cannot be run in read-write transactions
-        # hence this method exists to circumvent that limit.
-        self.run_prior_DDL_statements()
-
-        with self._database.snapshot() as snapshot:
-            res = snapshot.execute_sql(
-                sql, params=params, param_types=param_types
-            )
-            return list(res)
-
-    def get_table_column_schema(self, table_name):
-        rows = self.run_sql_in_snapshot(
-            """SELECT
-                COLUMN_NAME, IS_NULLABLE, SPANNER_TYPE
-            FROM
-                INFORMATION_SCHEMA.COLUMNS
-            WHERE
-                TABLE_SCHEMA = ''
-            AND
-                TABLE_NAME = @table_name""",
-            params={"table_name": table_name},
-            param_types={"table_name": spanner.param_types.STRING},
-        )
-
-        column_details = {}
-        for column_name, is_nullable, spanner_type in rows:
-            column_details[column_name] = ColumnDetails(
-                null_ok=is_nullable == "YES", spanner_type=spanner_type
-            )
-        return column_details
-
     def close(self):
-        """Close this connection.
+        """Closes this connection.
 
         The connection will be unusable from this point forward.
         """
-        self.rollback()
         self._database = None
         self.is_closed = True
 
     def commit(self):
+        """Commits any pending transaction to the database."""
         self._raise_if_closed()
 
         self.run_prior_DDL_statements()
@@ -142,9 +60,81 @@ class Connection(object):
         """A no-op, raising an error if the connection is closed."""
         self._raise_if_closed()
 
+    def cursor(self):
+        """Factory to create a DB-API Cursor."""
+        self._raise_if_closed()
+
+        return Cursor(self)
+
+    def run_prior_DDL_statements(self):
+        self._raise_if_closed()
+
+        if not self.ddl_statements:
+            return
+
+        ddl_statements = self.ddl_statements
+        self.ddl_statements = []
+
+        return self._database.update_ddl(ddl_statements).result()
+
     def __enter__(self):
         return self
 
     def __exit__(self, etype, value, traceback):
         self.commit()
         self.close()
+
+
+# def connect(
+#     instance_id, database_id, project=None, credentials=None, user_agent=None
+# ):
+#     """Creates a connection to a Google Cloud Spanner database.
+#
+#     :type instance_id: str
+#     :param instance_id: ID of the instance to connect to.
+#
+#     :type database_id: str
+#     :param database_id: The name of the database to connect to.
+#
+#     :type project: str
+#     :param project: (Optional) The ID of the project which owns the
+#                     instances, tables and data. If not provided, will
+#                     attempt to determine from the environment.
+#
+#     :type credentials: :class:`~google.auth.credentials.Credentials`
+#     :param credentials: (Optional) The authorization credentials to attach to
+#                         requests. These credentials identify this application
+#                         to the service. If none are specified, the client will
+#                         attempt to ascertain the credentials from the
+#                         environment.
+#
+#     :type user_agent: str
+#     :param user_agent: (Optional) Prefix to the user agent header.
+#
+#     :rtype: :class:`google.cloud.spanner_dbapi.connection.Connection`
+#     :returns: Connection object associated with the given Google Cloud Spanner
+#               resource.
+#
+#     :raises: :class:`ValueError` in case of given instance/database
+#              doesn't exist.
+#     """
+#     client_info = google_client_info(user_agent)
+#
+#     client = spanner_v1.Client(
+#         project=project,
+#         credentials=credentials,
+#         # client_info=google_client_info(user_agent),
+#         client_info=client_info,
+#     )
+#
+#     instance = client.instance(instance_id)
+#     if not instance.exists():
+#         raise ValueError("instance '%s' does not exist." % instance_id)
+#
+#     database = instance.database(
+#         database_id, pool=spanner_v1.pool.BurstyPool()
+#     )
+#     if not database.exists():
+#         raise ValueError("database '%s' does not exist." % database_id)
+#
+#     return Connection(database)
